Thread authorization policy added

+++ app/Http/Controllers/ThreadController.php
@@ -107,7 +107,8 @@ class ThreadController extends Controller
      */
     public function destroy(Channel $channel, Thread $thread)
     {
-        $thread->replies()->delete();
+        $this->authorize('delete', $thread);
+
         $thread->delete();
 
         return redirect('/threads');
